Rewrite the README
This commit rewrites the README for this project.

The author has been going back and fourth between whether to try and
find a replacement for this project or not; cookiecutter being the most
promising alternative.

However, the ease of installation and use of this project is extremely
compelling.

Given this, the author is choosing instead to take on some level of
burden to fix its shortcomings, rather than attempt to redo things in
another tool.

Further, this tool is written in the go language -- one of the primary
languages of LMC.