Refactor the application to use git mod, repo structure
Currently the repositiory is using "glide" for go version dependency.
The standard with more recent go projects is "go module".

This commit modifies the project to use the go module style.

== Design Notes
=== Package Name

Because the package is to be maintained by littleman.co going forward,
it uses the littleman.co package name spec.

=== Renaming the project

This commit also continues the project of renaming the project "boilr" →
"oilr". Future work will continue this process as references to boilr
become more apparent.